+++ src/util/patcher/index.js
@@ -1,0 +1,9 @@
+export { inject, uninject, isInjected } from './base';
+
+import * as _contextMenu from './contextMenu';
+
+export const contextMenu = _contextMenu;
+
+export const setThisScope = (scope) => {
+  _contextMenu.setThisScope(scope);
+};